Add scripts to convert OCDS releases to BigQuery format.
Currently this script *manually* removes extra columns not in the schema -
this was a quick fix and also useful to check how common it was for publishers
to include these (it turns out, pretty common).

This script should be amended to do this automatically instead, working against
a supplied schema.


Former-commit-id: e40b845c8efcc4ac4ca8cb1cc5ed663b01e8b57b